feature: add adapter popup and link button
