+++ apps/ai-dial-admin/src/components/EntityMainProperties/AdditionalProperties.tsx
@@ -5,7 +5,7 @@ import { FC, useCallback, useEffect, useRef, useState } from 'react';
 import { getModelsAdapters } from '@/src/app/[lang]/models/actions';
 import { TextInputField } from '@/src/components/Common/InputField/InputField';
 import ReadonlyField from '@/src/components/Common/ReadonlyField/ReadonlyField';
-import { CreateI18nKey, EntitiesI18nKey } from '@/src/constants/i18n';
+import { ButtonsI18nKey, CreateI18nKey, EntitiesI18nKey } from '@/src/constants/i18n';
 import { useNotification } from '@/src/context/NotificationContext';
 import { useI18n } from '@/src/locales/client';
 import { DialAdapter } from '@/src/models/dial/adapter';
@@ -76,6 +76,8 @@ const AdditionalProperties: FC<Props> = ({ view, entity, runners, onChangeEntity
 
       {view === ApplicationRoute.Models ? (
         <SourceEntitySelector
+          route={ApplicationRoute.Adapters}
+          buttonTitle={t(ButtonsI18nKey.OpenAdapter)}
           columns={SIMPLE_DESCRIPTION_COLUMNS}
           selectedValue={entity.adapter}
           fieldTitle={t(CreateI18nKey.AdapterTitle)}
